Move content from legacy and delete
